Made Album an aggregate that owns Track instances. Added album length and prettified lengths.

+++ src/infrastructure/musicbrainz/media_track.rs
@@ -3,7 +3,7 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct MediaTrack {
-    pub length: Option<u32>,
+    pub length: Option<u64>,
     pub id: String,
     pub number: String,
     pub title: String,
